Centralize link settings panel for button and link blocks
